third_party/WebKit/LayoutTests/fast/events/touch/gesture/touch-gesture-scroll-div-past-extent-diagonally-expected.txt

This tests that a gesture scroll isn't propagated from an inner div to an outer div when the inner div has remaining scroll offset on one axis but not on the other, unless the outer div starts at its scroll extent

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
PASS horizontal.scrollLeft is 15
PASS vertical.scrollTop is 0
PASS horizontal.scrollLeft is 600
PASS vertical.scrollTop is 0
PASS horizontal.scrollLeft is 600
PASS vertical.scrollTop is 20
